Pos.Util.Trace module defined
Had to shuffle the Pos.Util.Orphans module around: it defined Mockable
stuff which was dependent upon cardano-sl-networking. My solution is
still awful but the plan is to gradually dump Mockable anyway.
