+++ networking/cardano-sl-networking.cabal
@@ -65,6 +65,7 @@ Library
                       , async
                       , attoparsec
                       , base
+                      , cardano-sl-util
                       , containers
                       , contravariant
                       , cryptonite
